Use random non-inline symbols in `mrb_ivbm_create_symbols`
The current hash function for inline symbols is biased. Therefore, if
inline symbols whose symbol names are serial numbers are used, the value
that can be taken as the index of the hash bucket is limited. Then, use the
non-inline symbols. However, if they are just serial numbers, collisions
will rarely occur and it is not realistic, so random values are used.